Separate classes into individual files
