+++ osu.Game/Screens/OnlinePlay/Matchmaking/Queue/ScreenQueue.cs
@@ -17,7 +17,6 @@ using osu.Framework.Extensions.ObjectExtensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Audio;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Effects;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input.Bindings;
@@ -30,7 +29,6 @@ using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Input.Bindings;
-using osu.Game.Localisation;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Online.Chat;
@@ -39,15 +37,12 @@ using osu.Game.Online.Matchmaking.Requests;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Multiplayer.MatchTypes.RankedPlay;
 using osu.Game.Overlays;
-using osu.Game.Overlays.Mods;
 using osu.Game.Overlays.Volume;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Screens.Footer;
 using osu.Game.Screens.OnlinePlay.Matchmaking.Match;
 using osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay;
-using osu.Game.Screens.Play.HUD;
-using osu.Game.Screens.Select;
 using osuTK;
 using osuTK.Graphics;
 
@@ -955,144 +950,6 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Queue
                 base.Update();
 
                 Text = queue.QueueTimer.Elapsed.ToString(@"mm\:ss");
-            }
-        }
-
-        private partial class RankedPlayModSelectOverlay : UserModSelectOverlay
-        {
-            public new Func<Mod, bool> IsValidMod
-            {
-                get => base.IsValidMod;
-                set => base.IsValidMod = m => m.UserPlayable && value.Invoke(m);
-            }
-
-            public RankedPlayModSelectOverlay()
-                : base(OverlayColourScheme.Plum)
-            {
-                IsValidMod = _ => true;
-            }
-
-            public override VisibilityContainer CreateFooterContent() => new RankedPlayModSelectFooterContent(this)
-            {
-                Beatmap = { BindTarget = Beatmap },
-                ActiveMods = { BindTarget = ActiveMods },
-                Ruleset = { BindTarget = Ruleset },
-            };
-
-            public partial class RankedPlayModSelectFooterContent : ModSelectFooterContent
-            {
-                protected override bool ShowModEffects => false;
-
-                public RankedPlayModSelectFooterContent(RankedPlayModSelectOverlay overlay)
-                    : base(overlay)
-                {
-                }
-
-                protected override IEnumerable<ShearedButton> CreateButtons() => [];
-            }
-        }
-
-        private partial class RankedPlayFooterButtonFreeMods : ScreenFooterButton
-        {
-            public new readonly IBindable<MatchmakingScreenState> State = new Bindable<MatchmakingScreenState>();
-
-            public readonly Bindable<IReadOnlyList<Mod>> Mods = new Bindable<IReadOnlyList<Mod>>([]);
-
-            public new Action Action
-            {
-                set => throw new NotSupportedException("The click action is handled by the button itself.");
-            }
-
-            [Resolved]
-            private OsuColour colours { get; set; } = null!;
-
-            [Resolved]
-            private OverlayColourProvider colourProvider { get; set; } = null!;
-
-            private Container modsWedge = null!;
-
-            public RankedPlayFooterButtonFreeMods(ModSelectOverlay overlay)
-                : base(overlay)
-            {
-            }
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                Text = OnlinePlayStrings.FooterButtonFreemods;
-                TooltipText = MultiplayerMatchStrings.FreeModsButtonTooltip;
-                Icon = FontAwesome.Solid.ExchangeAlt;
-                AccentColour = colours.Lime1;
-
-                Add(modsWedge = new InputBlockingContainer
-                {
-                    Y = -5f,
-                    Depth = float.MaxValue,
-                    Origin = Anchor.BottomLeft,
-                    Shear = OsuGame.SHEAR,
-                    CornerRadius = CORNER_RADIUS,
-                    Size = new Vector2(BUTTON_WIDTH, FooterButtonMods.BAR_HEIGHT),
-                    Masking = true,
-                    EdgeEffect = new EdgeEffectParameters
-                    {
-                        Type = EdgeEffectType.Shadow,
-                        Radius = 4,
-                        // Figma says 50% opacity, but it does not match up visually if taken at face value, and looks bad.
-                        Colour = Colour4.Black.Opacity(0.25f),
-                        Offset = new Vector2(0, 2),
-                    },
-                    Alpha = 0,
-                    Children = new Drawable[]
-                    {
-                        new Box
-                        {
-                            Colour = colourProvider.Background4,
-                            RelativeSizeAxes = Axes.Both,
-                        },
-                        new Container
-                        {
-                            CornerRadius = CORNER_RADIUS,
-                            RelativeSizeAxes = Axes.Both,
-                            Masking = true,
-                            Children = new Drawable[]
-                            {
-                                new ModDisplay(showExtendedInformation: true)
-                                {
-                                    Anchor = Anchor.Centre,
-                                    Origin = Anchor.Centre,
-                                    Shear = -OsuGame.SHEAR,
-                                    Scale = new Vector2(0.5f),
-                                    Current = { BindTarget = Mods },
-                                    ExpansionMode = ExpansionMode.AlwaysContracted,
-                                },
-                            }
-                        },
-                    }
-                });
-            }
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                State.BindValueChanged(s =>
-                {
-                    if (s.NewValue == MatchmakingScreenState.Idle)
-                        Enabled.Value = true;
-                    else
-                    {
-                        Enabled.Value = false;
-                        Overlay?.Hide();
-                    }
-                }, true);
-
-                Mods.BindValueChanged(m =>
-                {
-                    if (m.NewValue.Count == 0)
-                        modsWedge.FadeOut(300, Easing.OutExpo);
-                    else
-                        modsWedge.FadeIn(300, Easing.OutExpo);
-                }, true);
             }
         }
     }
